Move slugify filter to an adapter (saves ~30KB on coremin bundle)
