+++ src/Adapters/Configuration/Filters/Slugify.coremin.js
@@ -1,0 +1,5 @@
+export default function () {
+	throw new Error(
+		"The `slugify` filter not included with the minimal Eleventy bundle. You can add it yourself via `eleventyConfig.addFilter`",
+	);
+}
