+++ src/Adapters/Configuration/Filters/Slugify.js
